商品削除機能の修正

+++ app/controllers/items_controller.rb
@@ -36,12 +36,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if current_user.id == @item.user.id
+    if user_signed_in? && current_user.id == @item.user.id
       if @item.destroy
          redirect_to root_path
       else
         render :show
       end
+    else
+      render :show
     end
   end
